Check option TR_DisableIProfilerThread instead of iprofiler thread presence
In `TR_Options::feLatePostProcess()` we want to disable the
`TR_UseIdleTime` option bit if we don't want to run with
an IProfiler thread. Currently this is done by checking the
presence of the Iprofiler thread. The problem is that the
IProfiler thread is created much later so our query will
always return `false`.
This commit replaces the check on the IProfiler thread with
a check on the `TR_DisableIProfilerThread` option bit.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>