Do not trust nullable arrayclass type in VP
Update isUnreliableSignatureType to check if the class
is an array class, and if its component class is value
type, and if the array class is not null-restricted array.
If so, the type is not reliable because it can either be a
nullable array class type or a null-restricted array type.

Fixes: #19914
Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>