Extend CAS alignment error to Little Endian
compareAndExchange16BitsOffsetChecks is used by 16 bit compare and
exchange operations to check to see if the 16 bits cross a 32 bit
boundary which is not supported.

This check was previously only enabled under Big Endian but the problem
can occur under Little Endian as well.

The check has been updated to apply to Little Endian to prevent
incorrect behavior.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>