Fix codecache disclaim error checking
According to the docs, madvise will return -1 on error and set errno.
Fix the code to check errno for the error, rather than the return value.

The current code wants to disable codecache disclaim if a persistent
error occurs. Fix the code to do so if the error is anything other than
EAGAIN. If the error is EGAIN it's expected to be transient and we
should be able to continue to disclaim.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>