Enable inlining of non-native Unsafe unaligned getters and setters
For Unsafe_getXUnaligned methods, which are wrappers to
native methods that contain some runtime checks, we benefit from directly
inlining them in inlineUnsafeCall as if they were their underlying native
methods, if we can determine that it is safe to do so.

This commit introduces the changes necessary to correctly handle such
non-native unaligned getters and setters, and enables transforming them
to simple loads and stores based on whether the platform supports
aligned access only.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>