Update doPrivilegedWithCombinerHelper function
When we try to invoke doPrivilegedWithCombiner function to
perform a privileged action under an existing context
environment, we are used to construct a new context but ignore
the parent context.

We should take consideration of a combination of the current
and parent context, rather than just choose either the current
or the parent.

This patch solves a failed case in issue #19499.

Issue: #19499

Signed-off-by: Jinhang Zhang <Jinhang.Zhang@ibm.com>