Add a stub for java.lang.invoke.MethodHandleImpl
OpenJ9 JDK8 and JDK11 currently use OpenJ9 MethodHandles.

Enabling jdk_bean in JDK8 and JDK11 triggers certain OpenJDK
MethodHandles code paths. However, OpenJ9 MethodHandles are in use
for JDK8 and JDK11, so OpenJDK MethodHandles paths shouldn't be
activated. To address this, a MethodHandleImpl stub was added,
used only with OpenJ9 MethodHandles, and not when OpenJDK
MethodHandles are enabled.

Related: #20531

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>