Add storage, retrieval of AOT method dependencies
When the dependencies of compiled AOT method bodies are tracked, they
will be stored in the SCC as arrays of uintptr_t values, with the first
entry being the length of the array, and the subsequent entries being
encoded class chain offsets. The low bit of the class chain offset will
be cleared if the corresponding class merely needs to be loaded before
the compiled method body in the SCC can be loaded.

This infrastructure is currently unused.

Signed-off-by: Christian Despres <despresc@ibm.com>