Fix JFR synchronization issues
- Assert that VM access is always held while interacting with JFR buffer
- Previously, we stored classhashtable entries. Class hashtable can grow
and these entries become stale. Instead use romClass as the key for the
package table
- Previously, we iterated ClassEntry hashtables while adding entries to
them (see fixupShallowEntries), this results in corruption. Instead we
now make a list of the classes we need to scan (findShallowEntries),
then walk the list and add scan entries after.
- Dont create tables with J9HASH_TABLE_ALLOW_SIZE_OPTIMIZATION since we
cant use that flag and iterate them.

Other fixes:
- dont assume moduleName is always non-NULL
- skip thread flush if the buffer is NULL
- add aserts to non-sensible events

Signed-off-by: Tobi Ajila <atobia@ca.ibm.com>