Reset trampoline pointers to top of space
After flushing the code cache under FSD there is a call to resetTrampolines.
The _trampolineAllocationMark and _trampolineReservationMark need to be reset
to their starting positions but were incorrectly set to the end of the trampoline
space. As a result all active segments at the time of the cache flush are marked
as full (assuming a platform that uses trampolines). This can be seen by taking
a JIT verbose log with option `codecache` which will show

CODECACHE:  CodeCache 00007E759000C660 marked as full in reserveSpaceForTrampoline