Reset queued methods on cache flush
When running under FSD, class redefinition results in all methods being
unloaded from the class cache and the compilation queues are emptied.
This is partly accomplished by jitResetAllMethods in decomp.cpp, however
it misses those methods that have been queued but not yet compiled. The
result is that after class redef a subset of methods will never be
recompiled, with a potentially significant impact on performance. This
change ensures that queued methods also have their status reset.