Change J9::VMMethodEnv::startPC error return value
OMR version of startPC expects either a valid start PC or 0 to be
returned from J9::VMMethodEnv::startPC since it isn't aware of the J9
specific non-startPC special values that are currently being returned.

This change checks if the J9_STARTPC_NOT_TRANSLATED bit flag is set on
the startPC to be returned. This indicates that it isn't a real startPC
and will return 0 instead.

Error checking has also been added to addFreeBlock to handle startPC
returning 0.

Signed-off-by: jimmyk <jimmyk@ca.ibm.com>