Refine Symbol::isVolatile tests
With the expansion of possible memory ordering semantics from binary volatile
or non-volatile to volatile, acquire/release, opaque, and transparent, all
tests  whether a symbol is volatile need to be refined depending on the
intention of the test, i.e. is it testing if the symbol is strictly volatile,
simply opaque, or somewhere in between?

Signed-off-by: Spencer Comin <spencer.comin@ibm.com>