Modify compile-time static field ref resolution
The existing code was introduced to improve the relocation success rate
for AOT compilations that track dependencies, by attempting to resolve
static field refs in jitGetClassOfFieldFromCP if they were not already
resolved. This has introduced a regression in comp CPU for AOT stores,
so the attempted field resolution has now been moved to the relevant
record validation function to avoid this defect.

Signed-off-by: Christian Despres <despresc@ibm.com>