Add check for JFR enablement
Currently, JFR is enabled by default on all platforms that support it,
and it cannot be disabled. We have been using `-XX:+FlightRecorder` to
indicate if the JVM should start with recording enabled. This is not
consistent with the RI.

With this change `-XX:[+|-]FlightRecorder` toggles whether JFR is enabled.
By default JFR will remain enabled, but we now have the ability to
disable it. `-XX:StartFlightRecording` will be used to indicate that a
recording should be begin at startup.

Signed-off-by: tajila <atobia@ca.ibm.com>