Accelerate hasNegatives, countPositives on x86
Accelerate hasNegatives on x86 for JDk 9-18 and countPositives
on x86 for JDK 19+. This acceleration is incremental in that
it only delivers the desired performance boost for arrays of up
to 31 elements. Above that, it still outperforms default OpenJ9,
but larger SIMD instructions are needed to extend the maximum
performance boost beyond 31 elements. It also doesn't accelerate
hasNegatives for JDK 19+, thanks to a performance anomaly that
causes hasNegatives to perform significantly faster than expected
for short arrays when countPositives is inlined into it, to the
point that implementing my acceleration would slow it down

Signed-off-by: Dylan Tuttle <jdylantuttle@gmail.com>