Change DLT opt level based on option
Before this change one could force the optimization
level of a particular DLT compilation through a
command line option (dltOptLevel=...) only if the
TR_DebugDLT env var was defined. This was done to avoid
the overhead of searching the option subsets for a
given method.
This commit allows one to specify the optimization level
of DLT compilations without the need of an env var.
Examples:
  -Xjit:dltOptLevel=hot
  -Xjit:{myMethod}(dltOptLevel=cold)
The overhead for searching option sets has been eliminated
if no option subset uses the dltOptlevel= option

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>