Fix VM snapshot class loader initialization
The code for initializing the persisted applicationClassLoader and
extensionClassLoader from a VM snapshot was in an unreachable path.
On restore runs, neither of those class loaders will be null, but
there was a null check guarding the initialization.

This patch reorders some of those checks, and ensures that the
class loaders are initialized properly.

Signed-off-by: Nathan Henderson <nathan.henderson@ibm.com>