Iprofiler records the class for direct calls
The IProfiler does not profile direct calls.
However, the optimizer may create artificial profiler
entries when it calls setCallCount(). These
artificial entries are marked to be non-persistent
and we only set the frequency (totalCount) ignoring
the class of the method that is called.
This commit also records the class of the method to
be called. Doing so does not require additional storage.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>