Create an adequately sized J9Heap to store the UpcallThunk
Currently, the J9Heap size to store the UpcallThunk is restricted to
the page size. If the thunk is greater than the page size, then the
existing code ends in an infinite loop creating J9Heaps of page size.
The correct sized J9Heap will never be created so the code will never
progress.

To resolve this issue, an adequately sized J9Heap, aligned to the page
size, is created to store the UpcallThunk.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>