Code to periodically track IProfiler overhead
Every 5 seconds we will print the number of profiling entries,
the CPU taken by the IProfiler thread and the top 5 methods
that received profiling samples.
The code is not active, to minimize overhead. In order to
compile it in one needs to define `TRACK_SAMPLE_SOURCES`
Also, to activate it at runtime, one needs to define the
following env var: `TR_TrackIProfilerActivity`
The information appears in the verbose log if one is defined
or to stderr if there is no vlog.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>