Re-prioritize RCP tracepoints
Currently, the tracepoints for RCP memory allocations and classloading
are enabled by default. These are hot paths which means that we incur an
overhead when writing out the tracepoints. Additionally, in a large app
these tracepoints would dominate the trace output. For these reasons the
tracepoint levels should be increased.

Signed-off-by: tajila <atobia@ca.ibm.com>