Set min jump table entries when compiling interpreter
OpenXL on PPC has been known to have some issues with branch
tables/indirect branch prediction before, which, when compiling
OpenJ9 on AIX with OpenXL, resulted in a ~15% performance regression
from before (i.e.: when building with xlC). To eliminate this
degredation, the compiler flags "-mllvm -ppc-min-jump-table-entries=512"
can be used on the souce files containing the interpreter loop.

Signed-off-by: midronij <jackie.midroni@ibm.com>