Reduce the amount of fanin info collected by IProfiler
The only consumer of the fanin info is the inliner.
However, the inliner will ignore the fanin data
for any method (callee) that is less than 50 bytecodes.
Thus, it makes sense to avoid storing fanin info for
such small methods. With this change the number of
entries stored by the fanin mechanism is reduced
by about 72%. This is expected to reduce footprint
somewhat and to reduce the overhead of the
methodinfo hashtable which stores fanin info.
The threshold that decides what is considered a small
method from the fanin mechanism point of view can be
changed with `-Xjit:iprofilerFaninMethodMinSize=`
option (default is 50 bytecodes)

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>