Adjust inliner use of profiled fanin data
The inliner uses some heuristics that tries to avoid the
inlining of large callees that have already been compiled.
What constitutes a large method is determined in a
routine called `isLargeCompileMethod()`.
If the callee is in a large frequency block, then we always
want to inline that callee.
If the callee is in a medium or low frequency block, then we
look at callee size and at the fanin info (how many methods
call this callee). For low frequency blocks we are more
conservative with inlining than with medium frequency blocks.
The existing code contained a bug in that, if there was no fanin
info for a callee in a low/medium frequency block, we would always
inline the callee, regardless of its size.
This commit fixes this problem and considers the size of the callee
first and then the fanin info. If there is no fanin info, then
the size of the callee is the only criterion.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>