Effectively revert PR #20975
PR #20975 is known to cause some throughput regressions.
These regressions are caused by limiting the IProfiler
to collect fanin info only for methods larger than 50 bytecodes.
It turns out that the code in `isLargeCompiledMethod()` does
not refer to the size of the callee in the variable called
`bytecodeSize`. `bytecodeSize` is actually some computed value
which can be much bigger than the size of the callee.

This commit eliminates the code that restricts the fanin info
to the methods larger than 50 bytecodes, but leaves all the
other changes of PR #20975 intact.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>