Track JIT body invalidation reasons as bit flags
...and specify the reason whenever we invalidate a JIT body.

Post-restore exclusion was already specially distinguished by a flag in
TR_PersistentMethodInfo.

Preexistence invalidation should also have been similarly distinguished,
but the logic to avoid making new preexistence assumptions has been
treating all invalidations as potentially preexistence-related.

This change is in preparation to add a new invalidation reason: the
unloading of an inlined method. It will be important to know when a
method has had a JIT body invalidated due to unloading so that when
recompiling, inlining can be restricted to avoid repeated invalidation.