Add env vars to experiment with large method heuristic
The two callers of isLargeCompiledMethod appear to be inconsistent
with the "size" that gets passed in. In one case, the size is
the estimated call graph size, whereas in the other case, the size
is just the bytecode size. This commit changes isLargeCompiledMethod
such that the thresholds to compare against are refactored out and passed
in from the caller side, thus allowing different thresholds to be
used from the two callers. New env vars have been introduced to do some
experiments with passing in different size thresholds, though in
principle, we could also vary the other thresholds being passed in
at each caller.

Signed-off-by: Vijay Sundaresan <vijaysun@ca.ibm.com>