Reduce the number of VM_isStable messages
The frontend query `isStable(J9Class *fieldClass, int32_t cpIndex)`
determines if the field at given J9Class and cpIndex has the
@Stable annotation.
Before this commit, the JITServer implementation had to send a
VM_isStable message to the client for every such query.
This commit implements caching of the isStable values at JITServer.
Caching is done with a hashtable per J9Class, where the key is
the cpIndex we are interested in.
To further reduce the number of the hashtables that serve as a
cache, the `isStable()` will answer `false` for any non-bootstrap
class. The reason is that at the moment only bootstrap classs have
the @stable annotation. This behavior can be disabled by defining
the following environment variable:
`TR_DontIgnoreStableAnnotationForUserClasses`.
Another optimization regards the implementation of
`isArrayWithStableElements(cpIndex, owningMethod, comp)`. In here,
we first check to see if we are dealing with an array (which is
simpler) and only then call the isStable() query.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>