Build a graph of class loaders based on lifetime relationships
Each J9ClassLoader now maintains a set of class loaders known to live at
least as long at it (outlivingLoaders). When a class is looked up in a
loader L1, and the class that's found was defined by a different loader
L2, L1 adds the class to its hash table, guaranteeing that L2 will
outlive L1. Now when this occurs, L1 will additionally take note that L2
outlives it. In this way we construct a graph of loaders describing the
lifetime relationships that have been observed so far. This graph can be
searched to determine the full set of loaders that are known to outlive
any particular loader. It would be possible to skip reifying this graph
and to get the same result instead by scanning through the (potentially
much larger) class hash tables.

Additionally, identify the permanent loaders by getting them from
systemClassLoader, extensionClassLoader, and applicationClassLoader, but
only once they're fully initialized and frozen. It's possible for the
latter two to be replaced during bootstrapping.

Permanent loaders don't need an outliving loader set, since they're
outlived only by other permanent loaders, so instead outlivingLoaders is
set to a special value to indicate that the loader is permanent.

Permanence is propagated in that any loader that is found to outlive a
permanent loader must also be permanent. This will be taken into account
regardless of which is known first: the permanence or the lifetime
relationship.

Permanent loaders are reported to the JIT so that it can take advantage
of them when reasoning about class loader lifetimes. It is now possible
to get a vector of pointers to all permanent loaders known so far by
calling J9::Compilation::permanentLoaders().

At the moment, outlivingLoaders and permanentLoaders() are essentially
unused, but soon they will both be used for RetainedMethodSet.