Build a graph of class loaders based on lifetime relationships
Each J9ClassLoader now maintains a set of class loaders known to live at
least as long at it (outlivingLoaders). When a class is looked up in a
loader L1, and the class that's found was defined by a different loader
L2, L1 adds the class to its hash table, guaranteeing that L2 will
outlive L1. Now when this occurs, L1 will additionally take note that L2
outlives it. In this way we construct a graph of loaders describing the
lifetime relationships that have been observed so far. This graph can be
searched to determine the full set of loaders that are known to outlive
any particular loader. It would be possible to skip reifying this graph
and to get the same result instead by scanning through the (potentially
much larger) class hash tables.