Prevent JIT bodies from strictly outliving methods inlined into them
This is largely in preparation for constant references, but it should
also have benefits for code motion. For more detail, see the Motivation
section of the doc comment for OMR::RetainedMethodSet.

Much of this commit is J9::RetainedMethodSet, the OpenJ9 implementation
of OMR::RetainedMethodSet, which describes a set of methods that will
remain loaded under certain assumptions. It's represented as a set of
class loaders together with a set of anonymous classes, since most
classes are unloaded at class loader granularity, but anonymous classes
can be unloaded individually.

Given a particular method to assume, J9::RetainedMethodSet uses
J9ClassLoader::outlivingLoaders to determine which other loaders can be
inferred to remain loaded at least as long as the assumed method.

During inlining, each TR_CallSite and TR_CallTarget will now have an
associated RetainedMethodSet, which may be used to arrange to establish
certain lifetime relationships between the JIT body and inlined methods.
It can extend the lifetime of an inlined method (keepalive) or restrict
the lifetime of the JIT body (bond) to make sure that it won't run once
an inlined method has been unloaded. See the corresponding OMR commit
for more.

Keepalives are determined but not yet actually put into effect since we
don't yet have any way to create the necessary references. Again see the
OMR commit for more.

A bond is put into effect by creating an unload assumption so that when
the inlined method is unloaded, the JIT body will be invalidated in much
the same way as for preexistence. Once a JIT body has been invalidated
for this reason, any later recompilation of the same method will set the
JIT option dontInlineUnloadableMethods, which will cause the compiler to
refuse to inline any target that would require a bond. This way,
invalidation due to unloading can't cause more than one recompilation of
a given method.

In AOT compilations, the compile-time analysis is skipped because its
results can't be expected to hold at load time. Instead the analysis is
done and runtime assumptions are created as necessary during AOT load,
after all of the usual validations and relocations have succeeded, based
on the inlining table in the J9JITExceptionTable. There are no missing
entries because loads can only succeed if all inlined methods are found.

In addition to the keepalives mentioned above, the compilation object
has a separate set of keepalive classes to be used for cases in which
non-inlining transformations cause the IL to directly mention classes
(or data belonging to classes) that might not be guaranteed to exist for
the entire lifetime of the JIT body. This is used when refining linkTo*
and invokeBasic calls in the trees and when transformIndirectLoadChain()
transmutes a node into a class pointer loadaddr.