Delay evaluation of dataAddr node till OOL if ref count is 1
Having dataAddr alive across GC point can lead to a crash. DataAddr
pointer is evaluated into a object stack slot so GC expects it to have an
class pointer but it's just a pointer to first data element. To avoid this
we delay the evaluation of the dataAddr node to OOL.

Signed-off-by: Shubham Verma <shubhamv.sv@gmail.com>