Command line option to control CS relative headroom
Introduce -XXgc:concurrentScavengeAllocAverageBoost= option to control
headroom for CS. Similar to existing -Xgc:concurrentScavengeSlack=.

The former is a relative boost factor (to the existing allocation rate),
and the latter is an absoulte number in bytes.

Either one can be used to help with noisy applications with spikes in
allocations.

Also moving concurrentScavengeAllocDeviationBoost from -Xgc to -XXgc,
since both boost options are unlikely to be used with customers.

The values specified are integers that are 100x larger then the desired
values.

For example, 130 actually represents 1.30x boost factor.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>