Command line option to control CS relative headroom
Introduce -Xgc:concurrentScavengeAllocAverageBoost= option to control
headroom for CS. Similar to existing -Xgc:concurrentScavengeSlack=.

The former is a ralative boost factor (to the existing allocation rate),
and the latter is an absoulte number in bytes.

Either one can be used to help with noisy applications with spikes in
allocations.

The values specified are integers that are 10x larger then the desired
values.

For example, 13 actually represents 1.3x boost factor.

Signed-off-by: Aleksandar Micic <Aleksandar_Micic@ca.ibm.com>