Disable javap -bootclasspath test in LockWordAlignment for JDK24+
Currently, the entire LockWordAlignment test suite is disabled to
suppress failure #20539.

It was observed that javap no longer overrides core JCL classes, such
as java/lang/Object, in JDK24+. This behavior is also present in the
RI.

As a result, the LockWordAlignment test suite is being re-enabled,
with only the affected sub-test impacted by the change in
javap -bootclasspath behavior being disabled.

We plan to reach out to OpenJDK developers to confirm whether this is
an intentional behavior change or a bug. Until then, #20539 will
remain open.

Related: #20539

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>