Narrow scope of TR::StackMemoryRegion during SPMD parallelization
TR_SPMDKernelParallelizer::processSPMDKernelLoopForSIMDize creates a
TR::StackMemoryRegion object that's used to allocate space for a
TR_HashTab.  However, while that object is still in scope, another
TR_HashTab - _loopDataType - that was allocated in an earlier
TR::StackMemoryRegion is used and potentially has more entries added to
it.  Once processSPMDKernelLoopForSIMDize's TR::StackMemoryRegion goes
out of scope and then comes back into scope, the contents of
_loopDataType could become corrupted.

Fixed this by narrowing the scope of the TR::StackMemoryRegion object in
processSPMDKernelLoopForSIMDize to ensure it goes out of scope before
_loopDataType might be manipulated again.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>