Suppress inlining identityHashcode and fastIdentityHashcode
For helper methods identityHashcode and fastIdentityHashcode,
it may be possible to evaluate their results at compile
time when their enclosing method Object.hashCode() has a known
class object info being passed as an arg which we can evaluate
during compile time. This is done in Value Propagation, but
only if we prevent the inlining of the identityHashcode and
fastIdentityHashcode helper methods.

Signed-off-by: Nazim Bhuiyan <nubhuiyan@ibm.com>