Print info about class unloading in verbose log
With this commit, if `-Xjit:verbose={classUnloading}` is specified on
the command line, the jitHookClassesUnload() will print to the verbose log
the number of classes being unloaded, as well as information about
the total number of classes loaded/unloaded up to the current time.
Moreover, if either of -Xjit:verbose={codecache} or -XX:+PrintCodeCache
or -Xjit:printCodeCacheUsage are specified, the JIT will also print
information about code cache occupancy, before and after the
class unloading cycle.
This information is supposed to help with code cache diagnostics of
applications that repeatedly load and unlaod classes.

Signed-off-by: Marius <mpirvu@ca.ibm.com>