Move location of incGlobalClassUnloadID()
incGlobalClassUnloadID() function is used to increment a counter
every time a class unload event happens. This is used by the
interpreter profiler to determine if it needs to re-examine the
validity of the entries it looks at. If there was no class unload
operation since the last time it looked at an entry, it doesn't
have to check again.
Currently, incGlobalClassUnloadID() is called in jitHookAnonClassesUnload()
and jitHookClassLoaderUnload(), but we only have to do it once
per class unload cycle. This commit moves the location of
incGlobalClassUnloadID() to jitHookClassesUnload() which is called
only once per class unload cycle.

Signed-off-by: Marius <mpirvu@ca.ibm.com>